pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "databento>=0.47.0",
  "loguru>=0.7.3",
  "matplotlib>=3.10.0",
  "polars>=1.19.0",
  "python-dotenv",
  "scikit-learn",
]

[tool.uv]
dev-dependencies = ["marimo>=0.10.9"]
